// web/src/shared/api/generated/model/integrationProviderResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IntegrationMode } from './integrationMode';

export interface IntegrationProviderResult {
  integration: string;
  provider: string;
  mode: IntegrationMode;
  isEnabled: boolean;
}